# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Beauty Clinic Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /patients:
    get:
      operationId: listPatients
      tags: [patients]
      parameters:
        - name: q
          in: query
          schema: { type: string }
        - name: page
          in: query
          schema: { type: integer }
        - name: limit
          in: query
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PatientList"
    post:
      operationId: createPatient
      tags: [patients]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatientInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Patient"
...
  /patients/{id}:
    get:
      operationId: getPatient
      tags: [patients]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Patient"
        "404":
          description: Not found
    put:
      operationId: updatePatient
      tags: [patients]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatientUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Patient"
        "404":
          description: Not found
    delete:
      operationId: deletePatient
      tags: [patients]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
  /patients/{id}/appointments:
    get:
      operationId: listPatientAppointments
      tags: [patients]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppointmentWithDetailsList"
  /patients/{id}/account-transactions:
    get:
      operationId: listPatientAccountTransactions
      tags: [patients]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PatientAccountTransaction"
    post:
      operationId: createPatientAccountTransaction
      tags: [patients]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatientAccountTransactionInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PatientAccountTransaction"
...
components:
  schemas:
...
    Patient:
      type: object
      required: [id, fileNumber, name, phone, accountBalance, createdAt]
      properties:
        id: { type: integer }
        fileNumber: { type: string }
        name: { type: string }
        phone: { type: string }
        email: { type: ["string", "null"] }
        birthdate: { type: ["string", "null"] }
        gender: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
        tier: { type: ["string", "null"] }
        accountBalance: { type: integer }
        referrerType: { type: ["string", "null"] }
        referrerId: { type: ["integer", "null"] }
        referrerRate: { type: ["integer", "null"] }
        referrerName: { type: ["string", "null"] }
        createdAt: { type: integer }
    PatientInput:
      type: object
      required: [fileNumber, name, phone]
      properties:
        fileNumber: { type: string }
        name: { type: string }
        phone: { type: string }
        email: { type: string }
        birthdate: { type: string }
        gender: { type: string }
        notes: { type: string }
        tier: { type: string }
        accountBalance: { type: number }
        referrerType: { type: string }
        referrerId: { type: number }
        referrerRate: { type: number }
    PatientUpdate:
      type: object
      properties:
        fileNumber: { type: string }
        name: { type: string }
        phone: { type: string }
        email: { type: ["string", "null"] }
        birthdate: { type: ["string", "null"] }
        gender: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
        tier: { type: ["string", "null"] }
        accountBalance: { type: integer }
        referrerType: { type: ["string", "null"] }
        referrerId: { type: ["integer", "null"] }
        referrerRate: { type: ["integer", "null"] }
    PatientAccountTransaction:
      type: object
      required: [id, patientId, amount, type, createdAt]
      properties:
        id: { type: integer }
        patientId: { type: integer }
        amount: { type: integer }
        type: { type: string }
        description: { type: ["string", "null"] }
        paymentId: { type: ["integer", "null"] }
        createdAt: { type: integer }
    PatientAccountTransactionInput:
      type: object
      required: [amount, type]
      properties:
        amount: { type: integer }
        type: { type: string }
        description: { type: string }
        paymentId: { type: integer }
    PatientList:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Patient"
        total: { type: integer }
        page: { type: integer }
        limit: { type: integer }

    # ── Service ───────────────────────────────────────────────────────────────
...
    AppointmentWithDetails:
      type: object
      required: [id, patientId, serviceId, scheduledAt, status, createdAt]
      properties:
        id: { type: integer }
        patientId: { type: integer }
        serviceId: { type: integer }
        staffId: { type: ["integer", "null"] }
        scheduledAt: { type: integer }
        status: { type: string }
        notes: { type: ["string", "null"] }
        price: { type: ["integer", "null"] }
        discountId: { type: ["integer", "null"] }
        originalPrice: { type: ["integer", "null"] }
        deposit: { type: ["integer", "null"] }
        createdAt: { type: integer }
        patientName: { type: ["string", "null"] }
        patientPhone: { type: ["string", "null"] }
        patientFileNumber: { type: ["string", "null"] }
        patientTier: { type: ["string", "null"] }
        serviceName: { type: ["string", "null"] }
        servicePrice: { type: ["integer", "null"] }
        staffName: { type: ["string", "null"] }
        unitsUsed: { type: ["integer", "null"] }
        priceMode: { type: ["string", "null"] }
        unitPrice: { type: ["integer", "null"] }
        unitLabel: { type: ["string", "null"] }
        serviceUnitCount: { type: ["integer", "null"] }
    AppointmentWithDetailsList:
      type: object
      required: [data, total]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/AppointmentWithDetails"
        total: { type: integer }
...